@@ -7,11 +7,15 @@ import { Observable } from 'rxjs';
 })
 export class ApiService {
   private _configUrl =
-    'https://pxnet2.stat.fi:443/PXWeb/api/v1/fi/Postinumeroalueittainen_avoin_tieto/2021/paavo_pxt_12f7.px';
+    'https://pxdata.stat.fi/PxWeb/api/v1/fi/Postinumeroalueittainen_avoin_tieto/2022/paavo_pxt_12f7.px';
 
   constructor(private http: HttpClient) {}
 
   getConfig(body): Observable<any> {
     return this.http.post(this._configUrl, body);
+  }
+
+  getVariablesAndPostCodes(): Observable<any> {
+    return this.http.get(this._configUrl);
   }
 }
